actions: Move part into zerver.lib.test_classes.
